+++ crates/spacewasm_util/src/file.rs
@@ -76,9 +76,7 @@ impl WasmStream for FileStream {
                 Ok(None)
             }
             Ok(n) => {
-                let m = unsafe {
-                    InnerVec::from_raw_parts(buf.as_mut_ptr(), buf.capacity() as u32, n as u32)
-                };
+                let m = unsafe { InnerVec::from_raw_parts(buf.as_mut_ptr(), buf.capacity(), n) };
 
                 self.n += n;
                 self.used.insert(buf.as_mut_ptr(), buf);
